Refactor the flower of osiris Update the mod's dependency Removed unused codes
