+++ src/main/java/com/mcgoodtime/productionline/tiles/SubTileFlowerOfOsiris.java
@@ -1,7 +1,8 @@
-package com.mcgoodtime.productionline.tiles.subtile.functional;
+package com.mcgoodtime.productionline.tiles;
 
 import vazkii.botania.api.subtile.SubTileFunctional;
 
 public class SubTileFlowerOfOsiris extends SubTileFunctional {
+
 
 }
