--- src/main/java/lib/LibBlockNames.java
@@ -1,5 +1,0 @@
-package lib;
-
-public class LibBlockNames {
-    public static final String THE_FLOWER_OF_OSIRIS = "theFlowerOfOsiris";
-}
